gamepreview: don't check if URL is valid
URL is initialized correctly by the parsers so there's no reason to check if the URL is correct. In addition if URL is not set, getUrl() returns an empty string which is not a correct URL.
This change simplifies the code and removes exception handling in some places  improving parsing time.